requirements.txt
opencv-python
deepface
langgraph
twilio
python-dotenv
numpy
tf-keras
